fix: rename use-interval to useInterval + minor fixes
