+++ services/orchest-webserver/client/src/jobs-view/hooks/usePollJobsStatus.tsx
@@ -1,15 +1,25 @@
 import { useJobsApi } from "@/api/jobs/useJobsApi";
-import { useInterval } from "@/hooks/use-interval";
+import { useAsync } from "@/hooks/useAsync";
+import { useInterval } from "@/hooks/useInterval";
 import { hasValue } from "@orchest/lib-utils";
+import React from "react";
 
-/**
- * Fetches the status of the jobs by polling.
- */
+/** Keeps job statuses updated in the jobs store by polling the back-end. */
 export const usePollJobsStatus = () => {
-  const isStoreLoaded = useJobsApi(
-    (state) => hasValue(state.projectUuid) && hasValue(state.jobs)
-  );
+  const isFetchingRef = React.useRef(false);
+  const isStoreLoaded = useJobsApi((state) => hasValue(state.jobs));
   const updateStatus = useJobsApi((state) => state.updateStatus);
+  const { run, error, status } = useAsync();
 
-  useInterval(updateStatus, !isStoreLoaded ? undefined : 1000);
+  isFetchingRef.current = status === "PENDING";
+
+  const poll = React.useCallback(() => {
+    if (!isFetchingRef.current) {
+      run(updateStatus());
+    }
+  }, [run, updateStatus]);
+
+  React.useEffect(() => console.log(error), [error]);
+
+  useInterval(poll, !isStoreLoaded ? undefined : 1000);
 };
